+++ ext/GeometryOpsProjExt/GeometryOpsProjExt.jl
@@ -1,0 +1,8 @@
+module GeometryOpsProjExt
+
+using GeometryOps, Proj
+
+include("reproject.jl")
+include("segmentize.jl")
+
+end
