+++ src/transformations/reproject.jl
@@ -6,6 +6,8 @@ export reproject
 # to another. It uses the `Proj` package for the transformation, but this could be 
 # moved to an extension if needed.
 
+# Note that the actual implementation is in the `GeometryOpsProjExt` extension module.
+
 # This works using the [`apply`](@ref) functionality.
 
 """
@@ -35,43 +37,19 @@ needed if it is not retreivable from the geometry with `GeoInterface.crs(geometr
 - `time`: the time for the coordinates. `Inf` by default.
 $APPLY_KEYWORDS
 """
-function reproject(geom;
-    source_crs=nothing, target_crs=nothing, transform=nothing, kw...
-)
-    if isnothing(transform)
-        if isnothing(source_crs) 
-            source_crs = if GI.trait(geom) isa Nothing && geom isa AbstractArray
-                GeoInterface.crs(first(geom))
-            else
-                GeoInterface.crs(geom)
-            end
-        end
+function reproject end
 
-        # If its still nothing, error
-        isnothing(source_crs) && throw(ArgumentError("geom has no crs attatched. Pass a `source_crs` keyword"))
+# ## Method error handling
+# We also inject a method error handler, which 
+# prints a suggestion if the Proj extension is not loaded.
 
-        # Otherwise reproject
-        reproject(geom, source_crs, target_crs; kw...)
-    else
-        reproject(geom, transform; kw...)
-    end
-end
-function reproject(geom, source_crs, target_crs;
-    time=Inf,
-    always_xy=true,
-    transform=Proj.Transformation(Proj.CRS(source_crs), Proj.CRS(target_crs); always_xy),
-    kw...
-)
-    reproject(geom, transform; time, target_crs, kw...)
-end
-function reproject(geom, transform::Proj.Transformation; time=Inf, target_crs=nothing, kw...)
-    if _is3d(geom)
-        return apply(PointTrait(), geom; crs=target_crs, kw...) do p
-            transform(GI.x(p), GI.y(p), GI.z(p))
-        end
-    else
-        return apply(PointTrait(), geom; crs=target_crs, kw...) do p
-            transform(GI.x(p), GI.y(p))
-        end
+function _reproject_error_hinter(io, exc, argtypes, kwargs)
+    if isnothing(Base.get_extension(GeometryOps, :GeometryOpsProjExt))
+        print(io, "\n\nThe `reproject` method requires the Proj.jl package to be explicitly loaded.\n")
+        print(io, "You can do this by simply typing ")
+        printstyled(io, "using Proj"; color = :cyan, bold = true)
+        println(io, " in your REPL, \nor otherwise loading Proj.jl via using or import.")
+    else # this is a more general error
+        nothing
     end
 end
