+++ test/transformations/simplify.jl
@@ -1,6 +1,7 @@
 import GeoJSON
 import JLD2
 import LibGEOS as LG
+import GeometryOps as GO
 
 datadir = realpath(joinpath(dirname(pathof(GO)), "../test/data"))
 @testset "RadialDistance and VisvalingamWhyatt" begin
